+++ src/runtime/Tests/WireFormatTests.m
@@ -113,10 +113,10 @@
 
 - (void) assertFieldsInOrder:(NSData*) data {
   PBCodedInputStream* input = [PBCodedInputStream streamWithData:data];
-  int32_t previousTag = 0;
+  long previousTag = 0;
 
   while (YES) {
-    int32_t tag = [input readTag];
+    long tag = [input readTag];
     if (tag == 0) {
       break;
     }
